requirements.txt
python-telegram-bot==21.1
python-dotenv==1.0.0
pillow==10.2.0 